More bugfixes for constant refresh main menu items
it was mapping rmsL and rmsR as floats. They are now just truncated to their equivalent uint32_t value. Also had to fix the map maximum input value. If you send the progress bar a value > 100, it bugs the whole display out. These changes fix the display not responding to the amp reset button press.